use echo_http_server fixture instead of httpbin.org requests
